refactor: move the leaving course prompt modal window with to a separate component and change the color of the exclamation mark icon
